Fix LogisticSwaps by updating ExpyXBnoji correctly
The reason for doing this is that during the for-loop between lines 94-110, the update of 'ExpyXBnoji' always lags behind the update of 'Binew' by 1 step. Therefore, the final 'ExpyXBnoji' corresponds to the value at t-1 step, not the t step. However, Binew has been updated by t steps. This code fix should also work in the case that ObjTemp < Fmin and the algorithm skips over the for-loop between lines 94-100.